third_party/WebKit/LayoutTests/fast/dom/xmlhttprequest-constructor-in-detached-document-expected.txt

CONSOLE MESSAGE: line 15: Uncaught NetworkError: A network error occurred.
Text for bug 25290: Crash when constructing XMLHttpRequest in a detached document.

PASS
